============================================================ =                   WARNING - MAJOR BUGFIX                 = ============================================================ Some rare events could affect the list of promoters (introducing a shift, leaving a no longer existing promoter in the list or not find a newly created promoter).
Details :
  * Small deletions containing OriC could create a shift in the list of promoters.
  * Small insertions that made the size of the genome grow from less than PROM_SIZE
    to at least PROM_SIZE could "miss" a newly created promoter.
  * Small deletions that made the genome shrink to less than PROM_SIZE could leave a "ghost" promoter in the list.   
  * When the genome size was exactly PROM_SIZE, the promoter search process wasn't done properly


I had tested the rearrangement processes very hard, they appear to be exempt of such problems.
I had not tested the point mutations and the indels as thoroughly as the rearrangement processes.
As these events are rare (and almost never happen with some sets of parameters) these errors managed to sneeked through the tests.

I added a complete assertion of RNA lists at every replication when on DEBUG mode.
The DEBUG mode is hence A LOT SLOWER than the "normal" mode but it's a price I'm willing to pay to avoid such problems!
We could think of using several debug modes but...

**********************************************************************
*  Most importantly, we should periodically run long tests in DEBUG  *
*  mode (maybe 10000 generations) on different sets of parameters to *
*  make sure everything is fine!                                     *
**********************************************************************




I changed / added functions that remove promoters or look for new promoters in ae_genetic_unit
and I added a great bunch of comments explaining precisely what functions do.
I spent quite some time trying to get these comments to show up properly in the doxygen doc so
they can be used as examples for further function documentation.
I also added a "collaboration diagram" for each class in the doc. They are far from perfect
(especially because we use a lot of ae_list), but they could help a newcomer a lot and could
also be used as a starting point for a hand made diagram.





git-svn-id: svn+ssh://svn.gforge.liris.cnrs.fr/svnroot/aevol/trunk@198 fea4e401-09a1-4513-ac43-d68ac1e0e106