Several changes/fixes/additions, large and small.
I added an option to start a population from an individual organisms, saved as a text file containing a sequence of 0's and 1's. It is done by "-o file.txt" command option on startup. I am interested in doing "synthetic biology" with our Aevol organisms by moving genes/promotes around and designing new organism, in post-processing, which then requires the ability to re-load the newly contracted genomes into Aevol. To help with this, at the end of each run, the best organisms will also be saved into a best_last_org.txt file. At this point, the feature does not support X11 or revol.  

I've added a aevol_default_run sub-directory into the test directory. It has a nice, full param.in file, links to both aevol and aevol_X11 executables and is just meant to serve as a good place to start things when you download a fresh version of Aevol and want to show it off to someone. It's been annoying to always have to go into the test directories, modify param.in, etc. This should just run and look pretty. 

With the help of Renaud, all the selection features should work properly for spacial case. Apparently, I was sorting the fitness of the neighbors in the wrong direction, causing sometimes the least fit organism to be the most likely to reproduce in the next generation. Sigh. 

Last but not least... There were some changes in versions 224 and 225 that were not exactly great. They primarily concerned secretion, and in some cases they inadvertently reverted sections of ae_population.cpp and ae_individual.cpp to a much older version. All that is fixed now. I do apologize for any confusion this may have caused, and recommend that if had downloaded and used any of the past two version for your experiments, you move onto the current, clean version 225. 

 


git-svn-id: svn+ssh://svn.gforge.liris.cnrs.fr/svnroot/aevol/trunk@226 fea4e401-09a1-4513-ac43-d68ac1e0e106