========== In TRUNK ========== BUGFIX
* In ae_simulation.cpp, in the 
ae_simulation::ae_simulation( char* backup_file_name, ... ), the constructor
for starting the run from a backup file

The backup, dump, and tree directories were assumed to be there already. 
When the directory was not present, the run would crash. 

This could be considered a feature, but as we are now often restarting runs
there is no reason to expect that those should be there already. So, I just 
added a couple of mkdirs. If the directories are indeed present, that should
not create a problem. 




git-svn-id: svn+ssh://svn.gforge.liris.cnrs.fr/svnroot/aevol/trunk@236 fea4e401-09a1-4513-ac43-d68ac1e0e106