========== In TRUNK ==========
NEW

Added a new feature, the plasmid swap. It's like sex, but a bit less 
fun... a bit like retrotransfer.

Anyway, currently, there is a plasmid donor and a plasmid recipient. If 
the "swap_GUs" parameter is set to true (SWAP_GUS in param.in file), 
every time there would be a plasmid transfer, the "donor" sends a 
plasmid as before, but the "recipient" does the same. By default, this 
parameter is set to false.




git-svn-id: svn+ssh://svn.gforge.liris.cnrs.fr/svnroot/aevol/trunk@246 fea4e401-09a1-4513-ac43-d68ac1e0e106