BUGFIX:
When the rearrangements were similarity-based, the segment length stored 
in the ae_mutation was sometimes wrong. The correct length could however
be recomputed with ae_mutation::segment_length, which was luckily used
by the post-treatment fixed_mutations.

Now the bug is fixed, the _length attribute of the ae_mutation should be 
correct in all cases.



git-svn-id: svn+ssh://svn.gforge.liris.cnrs.fr/svnroot/aevol/trunk@265 fea4e401-09a1-4513-ac43-d68ac1e0e106