================= IN TRUNK ====================
* Patch to work around Bug #300: "Missing length 
data in the tree files for the rearrangements".

In the NORMAL (rather than LIGHT) tree mode, 
all mutations should be recorded in the tree files, 
with all the details about their positions and 
lengths. However, for the rearrangements, the length 
is not written on disk. Hence when we run the
postprocessing program called fixed_mutations, 
it writes -1 for the length of all rearrangements.

In this commit I wrote a patch for the fixed_mutations
program, which works with the faulty tree files.
This patch is for people who have started a large
simulation campaign with this version of the code.

However it is a rather ugly solution because the
ae_mutation class remains bugged somehow: it asks for
the length when a mutation is reported, but then does
not write it on disk, hence the information is lost and
cannot be restored from the other informations in the 
class. Indeed, the length of the genetic unit before the 
event is needed and is not stored in the ae_mutation objects. 

In the patch I just wrote for fixed_mutations, I was able  
to work around this problem because the lineage file 
contains the initial genome, and hence the intial length
of the genetic units. As we replay each mutation, we have
the length of the genetic unit at each step.  

In a future commit (hopefully tonight), I will 
correct the bug properly by including the length
in the tree files, which involves changing ae_mutation::save
and ae_mutation::ae_mutation( gzFile backup_file ). 


* The changes in ae_dna.cpp are just minor changes
in the comments (corrections of obvious copy-paste errors).



git-svn-id: svn+ssh://svn.gforge.liris.cnrs.fr/svnroot/aevol/trunk@510 fea4e401-09a1-4513-ac43-d68ac1e0e106